feat: api 명세에 맞게 api 파일구조와 기본 코드연결 구현
