+++ src/pages/admin/MemberManagement/api/updateMemberRole.ts
@@ -1,0 +1,15 @@
+import { apiInstance } from '@/app/api/initInstance';
+import { handleAxiosError } from '@/shared/utils/handleAxiosError';
+import type { MemberRole } from '../types/member';
+
+export const updateMemberRole = async (
+  clubId: string,
+  profileId: number,
+  role: MemberRole,
+): Promise<void> => {
+  try {
+    await apiInstance.patch(`/clubs/${clubId}/members/${profileId}/role`, { role });
+  } catch (e: unknown) {
+    return handleAxiosError(e, '회원 역할 변경에 실패했습니다.');
+  }
+};
